Generalize further to support similar operations with CUSIP identifiers
